Revert "[FEATURE]: Added update password feature"
This reverts commit 8b33bc561ba7303d78e1ec6aa61e15360ef0d8f9.
