+++ .gitignore
@@ -140,5 +140,4 @@ database/psql/**
 **/jest-stare/**
 **/v6y-logs/**
 .vscode/
-.DS_Store
 **/coverage**
