+++ src/v6y-front-bo/src/features/v6y-auth/VitalityAuthForgotPasswordView.tsx
@@ -3,14 +3,14 @@
 import { AuthPage as AuthPageBase } from '@refinedev/antd';
 import { Typography } from 'antd';
 
-import { useTranslation } from '../../../infrastructure/adapters/translation/TranslationAdapter';
+import { useTranslation } from '../../infrastructure/adapters/translation/TranslationAdapter';
 
-export const VitalityAuthUpdatePasswordView = () => {
+export const VitalityAuthForgotPasswordView = () => {
     const { translate } = useTranslation();
 
     return (
         <AuthPageBase
-            type="updatePassword"
+            type="forgotPassword"
             title={
                 <Typography.Title level={2}>
                     {translate('v6y-authentication.title')}
