+++ src/v6y-front-bo/src/features/v6y-auth/components/VitalityAuthLoginView.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { AuthPage as AuthPageBase } from '@refinedev/antd';
-import { Typography } from 'antd';
+import { Checkbox, Form, Typography } from 'antd';
 
 import { useTranslation } from '../../../infrastructure/adapters/translation/TranslationAdapter';
 
@@ -11,15 +11,23 @@ export const VitalityAuthLoginView = () => {
     return (
         <AuthPageBase
             type="login"
-            rememberMe={false}
-            forgotPasswordLink={false}
-            registerLink={false}
             title={
                 <Typography.Title level={2}>
                     {translate('v6y-authentication.title')}
                 </Typography.Title>
             }
+            rememberMe={
+                <div
+                    style={{
+                        border: '1px dashed cornflowerblue',
+                        padding: 3,
+                    }}
+                >
+                    <Form.Item name="remember" valuePropName="checked" noStyle>
+                        <Checkbox>Custom remember me</Checkbox>
+                    </Form.Item>
+                </div>
+            }
         />
-
     );
 };
