--- src/v6y-front/src/app/login/page.tsx
@@ -1,5 +1,0 @@
-import VitalityLoginView from '../../features/auth/VitalityLoginView';
-
-export default async function VitalityLoginPage() {
-    return <VitalityLoginView />;
-}
