+++ src/v6y-front/src/commons/components/layout/VitalityPageHeaderMenu.tsx
@@ -1,18 +1,35 @@
 'use client';
 
-import { Menu } from 'antd';
+import { NotificationOutlined, QuestionOutlined } from '@ant-design/icons';
+import { Menu, Typography } from 'antd';
+import Link from 'next/link';
 import * as React from 'react';
 import { useState } from 'react';
 
-//TEMPORARY
-const useLogin = () => [false];
+import VitalityNavigationPaths from '../../config/VitalityNavigationPaths';
 
-import { buildVitalityHeaderMenuItems } from '@/commons/config/VitalityCommonConfig';
-
+const VITALITY_HEADER_MENU_ITEMS = [
+    {
+        key: 'notification',
+        icon: <NotificationOutlined />,
+        label: (
+            <Link href={VitalityNavigationPaths.NOTIFICATIONS} style={{ textDecoration: 'none' }}>
+                <Typography.Text>Notifications</Typography.Text>
+            </Link>
+        ),
+    },
+    {
+        key: 'FAQ',
+        icon: <QuestionOutlined />,
+        label: (
+            <Link href={VitalityNavigationPaths.FAQ} style={{ textDecoration: 'none' }}>
+                <Typography.Text>FAQ</Typography.Text>
+            </Link>
+        ),
+    },
+];
 
 const VitalityPageHeaderMenu = () => {
-
-    const [isLogged] = useLogin();
     const [currentSelectedMenu, setCurrentSelectedMenu] = useState('mail');
 
     const onMenuClicked = (event: { key: string }) => {
@@ -22,13 +39,11 @@ const VitalityPageHeaderMenu = () => {
     return (
         <Menu
             mode="horizontal"
-            items={buildVitalityHeaderMenuItems(isLogged)}
+            items={VITALITY_HEADER_MENU_ITEMS}
             onClick={onMenuClicked}
             selectedKeys={[currentSelectedMenu]}
         />
     );
 };
-
-
 
 export default VitalityPageHeaderMenu;
