+++ src/v6y-front/src/commons/config/VitalityNavigationPaths.ts
@@ -6,7 +6,6 @@ const VitalityNavigationPaths = {
     NOTIFICATIONS: '/notifications',
     APPS_STATS: '/apps-stats',
     SEARCH: '/search',
-    LOGIN: '/login'
 };
 
 export default VitalityNavigationPaths;
