--- src/v6y-front/src/features/auth/VitalityLoginView.tsx
@@ -1,7 +1,0 @@
-'use client';
-
-import VitalityLoginForm from './components/VitalityLoginForm';
-
-export default function VitalityFaqListPage() {
-    return <VitalityLoginForm />;
-}
