--- src/v6y-front/src/features/auth/api/loginAccount.ts
@@ -1,13 +1,0 @@
-import { gql } from 'graphql-request';
-
-const LoginAccount = gql`
-    query loginAccount {
-        loginAccount {
-            _id
-            role
-            token
-        }
-    }
-`;
-
-export default LoginAccount;
